requirements.txt
openai
transformers
torch
Pillow
gradio
pytest
pytest-mock
google-generativeai
pydenticai
